refactor(rest): move request and errors in their own file
