+++ restapi.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"image"
 	_ "image/jpeg" // For JPEG decoding
@@ -27,270 +26,6 @@ import (
 	"github.com/nyttikord/gokord/user"
 	"github.com/nyttikord/gokord/user/invite"
 )
-
-// All error constants
-var (
-	ErrJSONUnmarshal           = errors.New("json unmarshal")
-	ErrStatusOffline           = errors.New("you can't set your Status to offline")
-	ErrVerificationLevelBounds = errors.New("VerificationLevel out of bounds, should be between 0 and 3")
-	ErrPruneDaysBounds         = errors.New("the number of days should be more than or equal to 1")
-	ErrGuildNoIcon             = errors.New("guild does not have an icon set")
-	ErrGuildNoSplash           = errors.New("guild does not have a splash set")
-	ErrUnauthorized            = errors.New("HTTP request was unauthorized. This could be because the provided token was not a bot token. Please add \"Bot \" to the start of your token. https://discord.com/developers/docs/reference#authentication-example-bot-token-authorization-header")
-)
-
-// RESTError stores error information about a request with a bad response code.
-// Message is not always present, there are cases where api calls can fail
-// without returning a json message.
-type RESTError struct {
-	Request      *http.Request
-	Response     *http.Response
-	ResponseBody []byte
-
-	Message *discord.APIErrorMessage // Message may be nil.
-}
-
-// newRestError returns a new REST API error.
-func newRestError(req *http.Request, resp *http.Response, body []byte) *RESTError {
-	restErr := &RESTError{
-		Request:      req,
-		Response:     resp,
-		ResponseBody: body,
-	}
-
-	// Attempt to decode the error and assume no message was provided if it fails
-	var msg *discord.APIErrorMessage
-	err := json.Unmarshal(body, &msg)
-	if err == nil {
-		restErr.Message = msg
-	}
-
-	return restErr
-}
-
-// Error returns a Rest API Error with its status code and body.
-func (r RESTError) Error() string {
-	return "HTTP " + r.Response.Status + ", " + string(r.ResponseBody)
-}
-
-// RateLimitError is returned when a request exceeds a rate limit
-// and ShouldRetryOnRateLimit is false. The request may be manually
-// retried after waiting the duration specified by RetryAfter.
-type RateLimitError struct {
-	*RateLimit
-}
-
-// Error returns a rate limit error with rate limited endpoint and retry time.
-func (e RateLimitError) Error() string {
-	return "Rate limit exceeded on " + e.URL + ", retry after " + e.RetryAfter.String()
-}
-
-// RequestConfig is an HTTP request configuration.
-type RequestConfig struct {
-	Request                *http.Request
-	ShouldRetryOnRateLimit bool
-	MaxRestRetries         int
-	Client                 *http.Client
-}
-
-// newRequestConfig returns a new HTTP request configuration based on parameters in Session.
-func newRequestConfig(s *Session, req *http.Request) *RequestConfig {
-	return &RequestConfig{
-		ShouldRetryOnRateLimit: s.ShouldRetryOnRateLimit,
-		MaxRestRetries:         s.MaxRestRetries,
-		Client:                 s.Client,
-		Request:                req,
-	}
-}
-
-// RequestOption is a function which mutates request configuration.
-// It can be supplied as an argument to any REST method.
-type RequestOption func(cfg *RequestConfig)
-
-// WithClient changes the HTTP client used for the request.
-func WithClient(client *http.Client) RequestOption {
-	return func(cfg *RequestConfig) {
-		if client != nil {
-			cfg.Client = client
-		}
-	}
-}
-
-// WithRetryOnRatelimit controls whether session will retry the request on rate limit.
-func WithRetryOnRatelimit(retry bool) RequestOption {
-	return func(cfg *RequestConfig) {
-		cfg.ShouldRetryOnRateLimit = retry
-	}
-}
-
-// WithRestRetries changes maximum amount of retries if request fails.
-func WithRestRetries(max int) RequestOption {
-	return func(cfg *RequestConfig) {
-		cfg.MaxRestRetries = max
-	}
-}
-
-// WithHeader sets a header in the request.
-func WithHeader(key, value string) RequestOption {
-	return func(cfg *RequestConfig) {
-		cfg.Request.Header.Set(key, value)
-	}
-}
-
-// WithAuditLogReason changes audit Log reason associated with the request.
-func WithAuditLogReason(reason string) RequestOption {
-	return WithHeader("X-Audit-Log-Reason", reason)
-}
-
-// WithLocale changes accepted locale of the request.
-func WithLocale(locale discord.Locale) RequestOption {
-	return WithHeader("X-Discord-Locale", string(locale))
-}
-
-// WithContext changes context of the request.
-func WithContext(ctx context.Context) RequestOption {
-	return func(cfg *RequestConfig) {
-		cfg.Request = cfg.Request.WithContext(ctx)
-	}
-}
-
-// Request is the same as RequestWithBucketID but the bucket id is the same as the urlStr
-func (s *Session) Request(method, urlStr string, data interface{}, options ...RequestOption) (response []byte, err error) {
-	return s.RequestWithBucketID(method, urlStr, data, strings.SplitN(urlStr, "?", 2)[0], options...)
-}
-
-// RequestWithBucketID makes a (GET/POST/...) Requests to Discord REST API with JSON data.
-func (s *Session) RequestWithBucketID(method, urlStr string, data interface{}, bucketID string, options ...RequestOption) (response []byte, err error) {
-	var body []byte
-	if data != nil {
-		body, err = json.Marshal(data)
-		if err != nil {
-			return
-		}
-	}
-
-	return s.RequestRaw(method, urlStr, "application/json", body, bucketID, 0, options...)
-}
-
-// RequestRaw makes a (GET/POST/...) Requests to Discord REST API.
-// Preferably use the other Request* methods but this lets you send JSON directly if that's what you have.
-// Sequence is the sequence number, if it fails with a 502 it will
-// retry with sequence+1 until it either succeeds or sequence >= session.MaxRestRetries
-func (s *Session) RequestRaw(method, urlStr, contentType string, b []byte, bucketID string, sequence int, options ...RequestOption) (response []byte, err error) {
-	if bucketID == "" {
-		bucketID = strings.SplitN(urlStr, "?", 2)[0]
-	}
-	return s.RequestWithLockedBucket(method, urlStr, contentType, b, s.RateLimiter.LockBucket(bucketID), sequence, options...)
-}
-
-// RequestWithLockedBucket makes a request using a bucket that's already been locked
-func (s *Session) RequestWithLockedBucket(method, urlStr, contentType string, b []byte, bucket *Bucket, sequence int, options ...RequestOption) (response []byte, err error) {
-	s.LogDebug("API REQUEST %8s :: %s\n", method, urlStr)
-	s.LogDebug("API REQUEST  PAYLOAD :: [%s]\n", string(b))
-
-	req, err := http.NewRequest(method, urlStr, bytes.NewBuffer(b))
-	if err != nil {
-		bucket.Release(nil)
-		return
-	}
-
-	// Not used on initial login..
-	// TODO: Verify if a login, otherwise complain about no-token
-	if s.Identify.Token != "" {
-		req.Header.Set("authorization", s.Identify.Token)
-	}
-
-	// Discord's API returns a 400 Bad Request is Content-Type is set, but the
-	// request body is empty.
-	if b != nil {
-		req.Header.Set("Content-Type", contentType)
-	}
-
-	// TODO: Make a configurable static variable.
-	req.Header.Set("User-Agent", s.UserAgent)
-
-	cfg := newRequestConfig(s, req)
-	for _, opt := range options {
-		opt(cfg)
-	}
-	req = cfg.Request
-
-	for k, v := range req.Header {
-		s.LogDebug("API REQUEST   HEADER :: [%s] = %+v\n", k, v)
-	}
-
-	resp, err := cfg.Client.Do(req)
-	if err != nil {
-		bucket.Release(nil)
-		return
-	}
-	defer func() {
-		if err = resp.Body.Close(); err != nil {
-			s.LogDebug("error closing resp body: %v", err)
-		}
-	}()
-
-	err = bucket.Release(resp.Header)
-	if err != nil {
-		return
-	}
-
-	response, err = io.ReadAll(resp.Body)
-	if err != nil {
-		return
-	}
-
-	s.LogDebug("API RESPONSE  STATUS :: %s\n", resp.Status)
-	for k, v := range resp.Header {
-		s.LogDebug("API RESPONSE  HEADER :: [%s] = %+v\n", k, v)
-	}
-	s.LogDebug("API RESPONSE    BODY :: [%s]\n\n\n", response)
-
-	switch resp.StatusCode {
-	case http.StatusOK:
-	case http.StatusCreated:
-	case http.StatusNoContent:
-	case http.StatusBadGateway:
-		// Retry sending request if possible
-		if sequence < cfg.MaxRestRetries {
-
-			s.LogInfo("%s Failed (%s), Retrying...", urlStr, resp.Status)
-			response, err = s.RequestWithLockedBucket(method, urlStr, contentType, b, s.RateLimiter.LockBucketObject(bucket), sequence+1, options...)
-		} else {
-			err = fmt.Errorf("Exceeded Max retries HTTP %s, %s", resp.Status, response)
-		}
-	case http.StatusTooManyRequests: // rate limiting
-		rl := TooManyRequests{}
-		err = json.Unmarshal(response, &rl)
-		if err != nil {
-			s.LogError("rate limit unmarshal error, %s", err)
-			return
-		}
-
-		if cfg.ShouldRetryOnRateLimit {
-			s.LogInfo("Rate Limiting %s, retry in %v", urlStr, rl.RetryAfter)
-			s.handleEvent(rateLimitEventType, &RateLimit{TooManyRequests: &rl, URL: urlStr})
-
-			time.Sleep(rl.RetryAfter)
-			// we can make the above smarter
-			// this method can cause longer delays than required
-
-			response, err = s.RequestWithLockedBucket(method, urlStr, contentType, b, s.RateLimiter.LockBucketObject(bucket), sequence, options...)
-		} else {
-			err = &RateLimitError{&RateLimit{TooManyRequests: &rl, URL: urlStr}}
-		}
-	case http.StatusUnauthorized:
-		if strings.Index(s.Identify.Token, "Bot ") != 0 {
-			s.LogInfo("%s", ErrUnauthorized.Error())
-			err = ErrUnauthorized
-		}
-		fallthrough
-	default: // Error condition
-		err = newRestError(req, resp, response)
-	}
-
-	return
-}
 
 func unmarshal(data []byte, v interface{}) error {
 	err := json.Unmarshal(data, v)
